v1.40.22 Released!
Added new exchange Binance
Many small fixes
Subscribe to beta of Qt Trader exchange https://qttrader.com

+++ src/exchange/exchange_bitfinex.h
@@ -56,8 +56,8 @@ private:
 
     QByteArray lastTradesDateCache;
 
-    quint64 lastTradesDate;
-    quint32 tickerLastDate;
+    qint64 lastTradesDate;
+    qint64 tickerLastDate;
     quint64 lastHistoryId;
 
     QList<DepthItem>* depthAsks;
