+++ src/exchange/exchange_gocio.h
@@ -63,7 +63,7 @@ private:
     QTime authRequestTime;
 
     quint32 lastPriceDate;
-    quint32 lastTickerDate;
+    qint64 lastTickerDate;
     quint32 privateNonce;
 
     QByteArray lastDate;
